make Snapshot directory & file extension configurable
This let's users adjust the behaviour of abide's interaction with the filesystem,
exporting Snapshot Dir & Ext variables.

I've adjusted the LoadSnapshots function to directly set the internal allSnapshots
variable (this is the only way loadSnapshots was being used). This is intended as
a short-ish term stopgap, allowing users to adjust vars like SnapshotDir then call
LoadSnapshots() to have these changes take effect. I've added documentation of this
to the readme.md.

closes #37